add comprobations from instructions

+++ Compiler3D/src/app/core/models/ast/main/utils/DataType.ts
@@ -1,5 +1,5 @@
 export enum DataType{
-    DOUBLE = 'float',
+    FLOAT = 'float',
     INTEGER = 'int',
     CHAR = 'char',
     STRING = 'string',
